+++ interpreter.cpp
@@ -1068,7 +1068,7 @@ int32 interpreter::get_operation_value(card* pcard, int32 findex, int32 extraarg
 		}
 		return OPERATION_FAIL;
 	}
-	int32 result = lua_isinteger(current_state, -1) ? lua_tointeger(current_state, -1) : std::round(lua_tonumber(current_state, -1));
+	int32 result = static_cast<int32>(lua_isinteger(current_state, -1) ? lua_tointeger(current_state, -1) : std::round(lua_tonumber(current_state, -1)));
 	lua_pop(current_state, 1);
 	no_action--;
 	call_depth--;
@@ -1109,7 +1109,7 @@ int32 interpreter::get_operation_value(card* pcard, int32 findex, int32 extraarg
 		if(lua_isboolean(current_state, index))
 			return_value = lua_toboolean(current_state, index);
 		else
-			return_value = std::round(lua_tonumber(current_state, index));
+			return_value = static_cast<int32>(std::round(lua_tonumber(current_state, index)));
 		result->push_back(return_value);
 	}
 	lua_settop(current_state, stack_top);
@@ -1133,9 +1133,9 @@ int32 interpreter::get_function_value(int32 f, uint32 param_count) {
 		if(lua_isboolean(current_state, -1))
 			result = lua_toboolean(current_state, -1);
 		else if(lua_isinteger(current_state, -1))
-			result = lua_tointeger(current_state, -1);
+			result = static_cast<int32>(lua_tointeger(current_state, -1));
 		else
-			result = std::round(lua_tonumber(current_state, -1));
+			result = static_cast<int32>(std::round(lua_tonumber(current_state, -1)));
 		lua_pop(current_state, 1);
 		no_action--;
 		call_depth--;
@@ -1169,9 +1169,9 @@ int32 interpreter::get_function_value(int32 f, uint32 param_count, std::vector<i
 			if(lua_isboolean(current_state, index))
 				return_value = lua_toboolean(current_state, index);
 			else if(lua_isinteger(current_state, index))
-				return_value = lua_tointeger(current_state, index);
+				return_value = static_cast<int32>(lua_tointeger(current_state, index));
 			else
-				return_value = std::round(lua_tonumber(current_state, index));
+				return_value = static_cast<int32>(std::round(lua_tonumber(current_state, index)));
 			result->push_back(return_value);
 		}
 		lua_settop(current_state, stack_top);
@@ -1236,7 +1236,7 @@ int32 interpreter::call_coroutine(int32 f, uint32 param_count, uint32 * yield_va
 	if (result == 0) {
 		coroutines.erase(f);
 		if(yield_value)
-			*yield_value = lua_isboolean(rthread, -1) ? lua_toboolean(rthread, -1) : lua_tointeger(rthread, -1);
+			*yield_value = static_cast<int32>(lua_isboolean(rthread, -1) ? lua_toboolean(rthread, -1) : lua_tointeger(rthread, -1));
 		current_state = lua_state;
 		call_depth--;
 		if(call_depth == 0) {
